+++ MagickCore/paint.c
@@ -288,8 +288,8 @@ MagickExport MagickBooleanType FloodfillPaintImage(Image *image,
                 if (IsFuzzyEquivalencePixelInfo(&pixel,target) == invert)
                   break;
                 SetPixelGray(floodplane_image,QuantumRange,q);
-                p+=GetPixelChannels(image);
-                q+=GetPixelChannels(floodplane_image);
+                p+=(ptrdiff_t) GetPixelChannels(image);
+                q+=(ptrdiff_t) GetPixelChannels(floodplane_image);
               }
               status=SyncCacheViewAuthenticPixels(floodplane_view,exception);
               if (status == MagickFalse)
@@ -316,8 +316,8 @@ MagickExport MagickBooleanType FloodfillPaintImage(Image *image,
             GetPixelInfoPixel(image,p,&pixel);
             if (IsFuzzyEquivalencePixelInfo(&pixel,target) != invert)
               break;
-            p+=GetPixelChannels(image);
-            q+=GetPixelChannels(floodplane_image);
+            p+=(ptrdiff_t) GetPixelChannels(image);
+            q+=(ptrdiff_t) GetPixelChannels(floodplane_image);
           }
         }
       start=x;
@@ -371,8 +371,8 @@ MagickExport MagickBooleanType FloodfillPaintImage(Image *image,
               ((image->alpha_trait & BlendPixelTrait) != 0))
             SetPixelAlpha(image,(Quantum) fill_color.alpha,q);
         }
-      p+=GetPixelChannels(floodplane_image);
-      q+=GetPixelChannels(image);
+      p+=(ptrdiff_t) GetPixelChannels(floodplane_image);
+      q+=(ptrdiff_t) GetPixelChannels(image);
     }
     if (SyncCacheViewAuthenticPixels(image_view,exception) == MagickFalse)
       status=MagickFalse;
@@ -856,8 +856,8 @@ MagickExport Image *OilPaintImage(const Image *image,const double radius,
         SetPixelChannel(paint_image,channel,p[j*(ssize_t)
           GetPixelChannels(linear_image)+i],q);
       }
-      p+=GetPixelChannels(linear_image);
-      q+=GetPixelChannels(paint_image);
+      p+=(ptrdiff_t) GetPixelChannels(linear_image);
+      q+=(ptrdiff_t) GetPixelChannels(paint_image);
     }
     if (SyncCacheViewAuthenticPixels(paint_view,exception) == MagickFalse)
       status=MagickFalse;
@@ -1012,7 +1012,7 @@ MagickExport MagickBooleanType OpaquePaintImage(Image *image,
           if ((traits & UpdatePixelTrait) != 0)
             SetPixelAlpha(image,(Quantum) conform_fill.alpha,q);
         }
-      q+=GetPixelChannels(image);
+      q+=(ptrdiff_t) GetPixelChannels(image);
     }
     if (SyncCacheViewAuthenticPixels(image_view,exception) == MagickFalse)
       status=MagickFalse;
@@ -1140,7 +1140,7 @@ MagickExport MagickBooleanType TransparentPaintImage(Image *image,
       GetPixelInfoPixel(image,q,&pixel);
       if (IsFuzzyEquivalencePixelInfo(&pixel,target) != invert)
         SetPixelAlpha(image,opacity,q);
-      q+=GetPixelChannels(image);
+      q+=(ptrdiff_t) GetPixelChannels(image);
     }
     if (SyncCacheViewAuthenticPixels(image_view,exception) == MagickFalse)
       status=MagickFalse;
@@ -1275,7 +1275,7 @@ MagickExport MagickBooleanType TransparentPaintImageChroma(Image *image,
         MagickFalse;
       if (match != invert)
         SetPixelAlpha(image,opacity,q);
-      q+=GetPixelChannels(image);
+      q+=(ptrdiff_t) GetPixelChannels(image);
     }
     if (SyncCacheViewAuthenticPixels(image_view,exception) == MagickFalse)
       status=MagickFalse;
